Opsplitsing van klasse Rijtechniek per aparte vaardigheid.
