+++ src/Models/Rijtechniek/Rijtechniek.java
@@ -1,0 +1,12 @@
+package Models.Rijtechniek;
+
+public class Rijtechniek {
+    private int kerenID;
+    private Rijtechniek_Kijktechniek kijkTechniek;
+    private Rijtechniek_Koppeling koppeling;
+    private Rijtechniek_Parkeren parkeren;
+    private Rijtechniek_Remtechniek remTechniek;
+    private Rijtechniek_Schakeltechniek schakelTechniek;
+    private Rijtechniek_Stuurtechniek stuurTechniek;
+    private Rijtechniek_Zithouding zithouding; 
+}
