+++ src/Models/Rijtechniek/Rijtechniek_Balanceren.java
@@ -1,0 +1,21 @@
+package Models.Rijtechniek;
+
+/**
+ *
+ * @author Beau
+ */
+public class Rijtechniek_Balanceren {
+    
+    private int balanceren;
+
+    public int getBalanceren() {
+        return balanceren;
+    }
+
+    public void setBalanceren(int balanceren) {
+        this.balanceren = balanceren;
+    }
+    
+    
+    
+}
