+++ src/Models/Rijtechniek/Rijtechniek_Koppeling.java
@@ -1,0 +1,20 @@
+package Models.Rijtechniek;
+
+/**
+ *
+ * @author Beau
+ */
+public class Rijtechniek_Koppeling {
+    private int koppeling;
+
+    public int getKoppeling() {
+        return koppeling;
+    }
+
+    public void setKoppeling(int koppeling) {
+        this.koppeling = koppeling;
+    }
+    
+    
+    
+}
