+++ src/Models/Rijtechniek/Rijtechniek_Parkeren.java
@@ -1,0 +1,39 @@
+package Models.Rijtechniek;
+
+/**
+ *
+ * @author Beau
+ */
+public class Rijtechniek_Parkeren {
+    
+    private int parkerenTweeVoertuigen;
+    private int parkerenAchterVoertuigen;
+    private int parkerenLinks;
+
+    public int getParkerenTweeVoertuigen() {
+        return parkerenTweeVoertuigen;
+    }
+
+    public void setParkerenTweeVoertuigen(int parkerenTweeVoertuigen) {
+        this.parkerenTweeVoertuigen = parkerenTweeVoertuigen;
+    }
+
+    public int getParkerenAchterVoertuigen() {
+        return parkerenAchterVoertuigen;
+    }
+
+    public void setParkerenAchterVoertuigen(int parkerenAchterVoertuigen) {
+        this.parkerenAchterVoertuigen = parkerenAchterVoertuigen;
+    }
+
+    public int getParkerenLinks() {
+        return parkerenLinks;
+    }
+
+    public void setParkerenLinks(int parkerenLinks) {
+        this.parkerenLinks = parkerenLinks;
+    }
+    
+    
+    
+}
