+++ src/Models/Rijtechniek/Rijtechniek_Remtechniek.java
@@ -1,0 +1,20 @@
+package Models.Rijtechniek;
+
+/**
+ *
+ * @author Beau
+ */
+public class Rijtechniek_Remtechniek {
+    private int remTechniek;
+
+    public int getRemTechniek() {
+        return remTechniek;
+    }
+
+    public void setRemTechniek(int remTechniek) {
+        this.remTechniek = remTechniek;
+    }
+    
+    
+    
+}
