+++ src/Models/Rijtechniek/Rijtechniek_Schakeltechniek.java
@@ -1,0 +1,20 @@
+package Models.Rijtechniek;
+
+/**
+ *
+ * @author Beau
+ */
+public class Rijtechniek_Schakeltechniek {
+    private int schakelTechniek;
+
+    public int getSchakelTechniek() {
+        return schakelTechniek;
+    }
+
+    public void setSchakelTechniek(int schakelTechniek) {
+        this.schakelTechniek = schakelTechniek;
+    }
+    
+    
+    
+}
