+++ src/Models/Rijtechniek/Rijtechniek_Stuurtechniek.java
@@ -1,0 +1,18 @@
+
+package Models.Rijtechniek;
+
+/**
+ *
+ * @author Beau
+ */
+public class Rijtechniek_Stuurtechniek {
+    private int stuurTechniek;
+
+    public int getStuurTechniek() {
+        return stuurTechniek;
+    }
+
+    public void setStuurTechniek(int stuurTechniek) {
+        this.stuurTechniek = stuurTechniek;
+    }   
+}
