+++ src/Models/Rijtechniek/Rijtechniek_VertrekkenHandrem.java
@@ -1,0 +1,20 @@
+package Models.Rijtechniek;
+
+/**
+ *
+ * @author Beau
+ */
+public class Rijtechniek_VertrekkenHandrem {
+    
+    private int vertrekkenHandrem;
+
+    public int getVertrekkenHandrem() {
+        return vertrekkenHandrem;
+    }
+
+    public void setVertrekkenHandrem(int vertrekkenHandrem) {
+        this.vertrekkenHandrem = vertrekkenHandrem;
+    }
+   
+    
+}
