+++ src/Models/Rijtechniek/Rijtechniek_VertrekkenHelling.java
@@ -1,0 +1,20 @@
+package Models.Rijtechniek;
+
+/**
+ *
+ * @author Beau
+ */
+public class Rijtechniek_VertrekkenHelling {
+    private int vertrekkenHelling;
+
+    public int getVertrekkenHelling() {
+        return vertrekkenHelling;
+    }
+
+    public void setVertrekkenHelling(int vertrekkenHelling) {
+        this.vertrekkenHelling = vertrekkenHelling;
+    }
+    
+    
+    
+}
