+++ src/Models/Rijtechniek/Rijtechniek_VertrekkenVoetrem.java
@@ -1,0 +1,20 @@
+package Models.Rijtechniek;
+
+/**
+ *
+ * @author Beau
+ */
+public class Rijtechniek_VertrekkenVoetrem {
+    
+    private int vertrekkenVoetrem;
+
+    public int getVertrekkenVoetrem() {
+        return vertrekkenVoetrem;
+    }
+
+    public void setVertrekkenVoetrem(int vertrekkenVoetrem) {
+        this.vertrekkenVoetrem = vertrekkenVoetrem;
+    }
+   
+    
+}
