+++ src/Models/Rijtechniek/Rijtechniek_Zithouding.java
@@ -1,0 +1,55 @@
+package Models.Rijtechniek;
+
+/**
+ *
+ * @author Beau
+ */
+public class Rijtechniek_Zithouding {
+    private int zithouding;
+    private int zithoudingGordels;
+    private int zithoudingSpiegels;
+    private int zithoudingHandrem;
+    private int zithoudingAndere;
+
+    public int getZithouding() {
+        return zithouding;
+    }
+
+    public int getZithoudingGordels() {
+        return zithoudingGordels;
+    }
+
+    public int getZithoudingSpiegels() {
+        return zithoudingSpiegels;
+    }
+
+    public int getZithoudingHandrem() {
+        return zithoudingHandrem;
+    }
+
+    public int getZithoudingAndere() {
+        return zithoudingAndere;
+    }
+
+    public void setZithouding(int zithouding) {
+        this.zithouding = zithouding;
+    }
+
+    public void setZithoudingGordels(int zithoudingGordels) {
+        this.zithoudingGordels = zithoudingGordels;
+    }
+
+    public void setZithoudingSpiegels(int zithoudingSpiegels) {
+        this.zithoudingSpiegels = zithoudingSpiegels;
+    }
+
+    public void setZithoudingHandrem(int zithoudingHandrem) {
+        this.zithoudingHandrem = zithoudingHandrem;
+    }
+
+    public void setZithoudingAndere(int zithoudingAndere) {
+        this.zithoudingAndere = zithoudingAndere;
+    }
+    
+   
+}
